feat: create currency field.
- Create currency field with custom formatters.
- Add extension for this custom formatters
- Apply currency field on add expense view.

+++ BudgetTracker/Structs/ExpenseItem.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct ExpenseItem: Identifiable, Codable {
     var id = UUID()
-    var amount: Double
+    var amount: Decimal
     var reasonForExpense: String
     var createdDate: Date
     var updateDate: Date
